# contracts/orchestrator.openapi.yaml
openapi: 3.1.0
info:
  title: Athlos Orchestrator API
  version: 1.0.0
  description: |
    Athlos-specific orchestration API. Prepares divisions, calls Competition Engine, persists results and sends webhooks.
servers:
  - url: https://orchestrator.example.com
paths:
  /v1/events/{event_id}/generate-all-brackets:
    post:
      summary: Prepare divisions and enqueue/generate all brackets for an event
      operationId: generateAllBrackets
      parameters:
        - name: event_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateAllRequest'
      responses:
        '202':
          description: Accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenerateAllResponse'
  /v1/events/{event_id}/divisions/{division_id}/lock-bracket:
    post:
      summary: Lock a bracket to prevent regeneration
      operationId: lockBracket
      parameters:
        - name: event_id
          in: path
          required: true
          schema:
            type: string
        - name: division_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Bracket locked
          content:
            application/json:
              schema:
                type: object
                properties:
                  event_id:
                    type: string
                  division_id:
                    type: string
                  locked:
                    type: boolean
  /v1/events/{event_id}/divisions/{division_id}/bracket:
    get:
      summary: Get the current bracket
      operationId: getBracket
      parameters:
        - name: event_id
          in: path
          required: true
          schema:
            type: string
        - name: division_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Bracket data
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '404':
          description: Bracket not found
  /v1/events/{event_id}/divisions/{division_id}/preview-bracket:
    post:
      summary: Preview bracket generation with diff
      operationId: previewBracket
      parameters:
        - name: event_id
          in: path
          required: true
          schema:
            type: string
        - name: division_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                overrides:
                  type: object
                  additionalProperties: true
      responses:
        '200':
          description: Preview with diff
          content:
            application/json:
              schema:
                type: object
                properties:
                  event_id:
                    type: integer
                  division_id:
                    type: integer
                  old_bracket:
                    type: object
                    nullable: true
                  new_bracket:
                    type: object
                  diff:
                    type: object
components:
  schemas:
    Webhook:
      type: object
      required: [url, secret]
      properties:
        url:
          type: string
        secret:
          type: string
    GenerateAllRequest:
      type: object
      required: [webhook]
      properties:
        webhook:
          $ref: '#/components/schemas/Webhook'
        overrides:
          type: object
          additionalProperties: true
    GenerateAllResponse:
      type: object
      required: [event_id, divisions_prepared, jobs_enqueued, jobs_skipped]
      properties:
        event_id:
          type: string
        divisions_prepared:
          type: boolean
        jobs_enqueued:
          type: integer
        jobs_skipped:
          type: integer